admin/requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
supabase==1.2.0